Has done lots of improvments
Is so much I have fix so this commit not cover it all.

But move thing to new classes, fixed so menu work and added more logic and fixed so menu work whit both player have metadata and cache.

Have for now remove, so you can only store location, not player in cache.