Fixed so two buttons can´t collide
I also cleaned up old code when I change how buttons get registered.

Is some breaking changes I have to made, for make it reliable. So old getItemAt is deprecated and not work any more.

The problem with  buttons was if two item was exactly same, my api could not get right button (I think is rare issue, but think is best if I solve it).